Create dependency injection
